Fix a bug while matching rows belonging to buckets of size > 1
- If we have four rows; the first & second have the same email,
and the third & fourth have the same email. At this point, the
first and second will share a match key, as will the third and
fourth. Now if the first and third have the secondary email in
common, all four rows need to have the same match key.
- While matching the first and third rows, we solve this by
going through all rows in the same bucket as the third row
and changing all their match keys to that of the first row